extracted env_vars mod
