+++ Cargo.toml
@@ -18,4 +18,4 @@ opt-level = 3
 debug = true
 
 [profile.dev]
-opt-level = 2
+opt-level = 0
