+++ Dockerfile
@@ -104,8 +104,7 @@ RUN microdnf update -y \
     && microdnf install ca-certificates libcap shadow-utils -y \
     && rm -rf /var/cache/yum \
     && chmod -R 777 . \
-    && cp /work/logdna-agent /work/logdna-agent-no-cap \
-    && setcap "cap_dac_read_search+eip" /work/logdna-agent \
+    && ln -s /work/logdna-agent /work/logdna-agent-no-cap \
     && groupadd -g 5000 logdna \
     && useradd -u 5000 -g logdna logdna
