+++ Dockerfile.debian
@@ -77,8 +77,7 @@ RUN apt update -y \
                    gdbserver \
     && rm -rf /var/cache/apt \
     && chmod -R 777 . \
-    && cp /work/logdna-agent /work/logdna-agent-no-cap \
-    && setcap "cap_dac_read_search+eip" /work/logdna-agent \
+    && ln -s /work/logdna-agent /work/logdna-agent-no-cap \
     && groupadd -g 5000 logdna \
     && useradd -u 5000 -g logdna logdna
