+++ Dockerfile.debug
@@ -75,7 +75,7 @@ RUN apt update -y \
                    dnsutils vim curl procps net-tools gdbserver -y \
     && rm -rf /var/cache/apt \
     && chmod -R 777 . \
-    && setcap "cap_dac_read_search+eip" /work/logdna-agent \
+    && ln -s /work/logdna-agent /work/logdna-agent-no-cap \
     && groupadd -g 5000 logdna \
     && useradd -u 5000 -g logdna logdna
