+++ common/config/Cargo.toml
@@ -7,7 +7,6 @@ edition = "2018"
 [dependencies]
 #local
 fs = { package = "fs", path = "../fs" }
-k8s = { package = "k8s", path = "../k8s" }
 http = { package = "http", path = "../http" }
 
 serde = { version = "1.0", features = ["derive"] }
@@ -15,6 +14,7 @@ serde_yaml = "0.8"
 globber = "0.1"
 lazy_static = "1.0"
 async-compression = "0.3"
+thiserror = "1.0"
 log = "0.4"
 sysinfo = "0.15"
 structopt = "0.3"
