+++ common/middleware/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 [dependencies]
 #local
 http = { package = "http", path = "../http" }
+config = { package = "config", path = "../config" }
 memoffset = "0.6"
 regex = "1"
 thiserror = "1.0"
